Atividade4 - corrigida

+++ AtividadesJava/src/Atividade4.java
@@ -1,6 +1,6 @@
 public class Atividade4 {
     public static void main(String[] args) {
-        final double PI = 3.1415; int raio = 5;
+        final double PI = 3.1415; double raio = 5;
         System.out.println("PI = "+PI);
         System.out.println("Raio = "+raio);
     }
